# specs/001-shared-subscription-spec/feature-admin-create-user/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Admin Creates Users API
  version: 1.0.0
  description: API contracts to support admin-driven user creation and invite link provisioning.
servers:
  - url: https://api.example.com
paths:
  /users:
    post:
      summary: Create a new user (admin only)
      tags: [Users]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, name]
              properties:
                email:
                  type: string
                  format: email
                name:
                  type: string
                role:
                  type: string
                  enum: [ADMIN, USER]
                  default: USER
                is_active:
                  type: boolean
                  default: true
      responses:
        '201':
          description: User created and setup link generated
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: '#/components/schemas/User'
                  setup_link:
                    type: string
                    description: One-time password setup link; displayed once to admin
        '400':
          description: Validation error
        '401':
          description: Unauthorized
        '403':
          description: Forbidden (admin only)
        '409':
          description: Email already exists

  /users/{id}/invitation:
    post:
      summary: Regenerate a user password setup link (admin only)
      tags: [Users]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: New one-time setup link generated; previous invalidated
          content:
            application/json:
              schema:
                type: object
                properties:
                  setup_link:
                    type: string
        '404':
          description: User not found

  /auth/setup-password:
    post:
      summary: Complete account setup by setting password using a one-time token
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token, password]
              properties:
                token:
                  type: string
                  description: One-time setup token
                password:
                  type: string
                  format: password
      responses:
        '200':
          description: Password set and account activated
        '400':
          description: Invalid or expired token / weak password
        '409':
          description: Token already used

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
          format: email
        name:
          type: string
        role:
          type: string
          enum: [ADMIN, USER]
        is_active:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time